api/requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<6.0
Flask-SQLAlchemy>=3.1,<4.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
PyJWT>=2.8,<3.0
python-dotenv>=1.0,<2.0
